jwt/backend/requirements.txt
black
flask-restful
flask-cors
flask-jwt-extended
webargs